+++ src/useForm.ts
@@ -101,17 +101,21 @@ export default <V extends FormValues = FormValues>({
       Array.from(form.querySelectorAll("input,textarea,select"))
         .filter((element) => {
           const field = element as FieldElement;
-          const { type, name, dataset } = field;
+          const {
+            type,
+            name,
+            dataset: { rcfIgnore },
+          } = field;
 
           if (/image|submit|reset/.test(type)) return false;
-          if (!name) {
+          if (rcfIgnore && !name) {
             warn('💡 react-cool-form > field: Missing the "name" attribute.');
             return false;
           }
 
           return (
             controllersRef.current[name] ||
-            !(dataset.rcfIgnore || ignoreFieldsRef.current[name])
+            !(rcfIgnore || ignoreFieldsRef.current[name])
           );
         })
         .reduce((acc: Record<string, any>, cur) => {
@@ -399,9 +403,11 @@ export default <V extends FormValues = FormValues>({
 
       if (builtInValidationMode === "message") return field.validationMessage;
 
-      // @ts-expect-error
       // eslint-disable-next-line no-restricted-syntax
-      for (const k in field.validity) if (field.validity[k]) return k;
+      for (const k in field.validity) {
+        // @ts-expect-error
+        if (k !== "valid" && field.validity[k]) return k;
+      }
 
       return undefined;
     },